# Use ConfigCache as second-level cache