--- a/src/Tree/RouteTree.php
+++ b/src/Tree/RouteTree.php
@@ -5,41 +5,61 @@ declare(strict_types=1);
 namespace Becklyn\RouteTreeBundle\Tree;
 
 use Becklyn\Menu\Item\MenuItem;
-use Becklyn\RouteTreeBundle\Builder\ItemCollectionBuilder;
+use Becklyn\RouteTreeBundle\Builder\ItemCollection;
 use Becklyn\RouteTreeBundle\Cache\TreeCache;
 use Becklyn\RouteTreeBundle\Exception\InvalidRouteTreeException;
+use Becklyn\RouteTreeBundle\Node\ItemFactory;
+use Symfony\Component\Config\ConfigCacheFactory;
+use Symfony\Component\Config\ConfigCacheFactoryInterface;
+use Symfony\Component\Config\ConfigCacheInterface;
 use Symfony\Component\HttpKernel\CacheClearer\CacheClearerInterface;
 use Symfony\Component\HttpKernel\CacheWarmer\CacheWarmerInterface;
+use Symfony\Component\Routing\RouterInterface;
 
 /**
  *
  */
 class RouteTree implements CacheClearerInterface, CacheWarmerInterface
 {
-    /**
-     * @var MenuItem[]
-     */
+    private const CONFIG_CACHE_PATH = "becklyn/route_tree/tree.serialized";
+
+    /** @var MenuItem[] */
     private $items;
 
+    /** @var TreeCache */
+    private $treeCache;
+
+    /** @var ConfigCacheFactoryInterface */
+    private $configCacheFactory;
+
+    /** @var string */
+    private $cacheDir;
+
+    /** @var bool */
+    private $isDebug;
+
+    /** @var ItemFactory */
+    private $itemFactory;
+
+    /** @var RouterInterface */
+    private $router;
+
 
     /**
-     * @var ItemCollectionBuilder
      */
-    private $collectionBuilder;
-
-
-    /**
-     * @var TreeCache
-     */
-    private $cache;
-
-
-    /**
-     */
-    public function __construct (ItemCollectionBuilder $collectionBuilder, TreeCache $cache)
+    public function __construct (
+        RouterInterface $router,
+        ItemFactory $itemFactory,
+        TreeCache $treeCache,
+        string $cacheDir,
+        bool $isDebug
+    )
     {
-        $this->collectionBuilder = $collectionBuilder;
-        $this->cache = $cache;
+        $this->treeCache = $treeCache;
+        $this->cacheDir = $cacheDir;
+        $this->isDebug = $isDebug;
+        $this->itemFactory = $itemFactory;
+        $this->router = $router;
     }
 
 
@@ -50,13 +70,31 @@ class RouteTree implements CacheClearerInterface, CacheWarmerInterface
      */
     private function generateItems () : array
     {
-        $nodes = $this->cache->get();
+        $nodes = $this->treeCache->get();
 
         if (null === $nodes)
         {
-            $collection = $this->collectionBuilder->build();
-            $nodes = $collection->getItems();
-            $this->cache->set($nodes);
+            $configCache = $this->getConfigCacheFactory()->cache(
+                "{$this->cacheDir}/" . self::CONFIG_CACHE_PATH,
+                function (ConfigCacheInterface $cache) : void
+                {
+                    $routeCollection = $this->router->getRouteCollection();
+
+                    $collection = new ItemCollection($this->itemFactory, $routeCollection);
+                    $items = $collection->getItems();
+
+                    $cache->write(
+                        \sprintf(
+                            '<?php return \unserialize(%s);',
+                            var_export(\serialize($items), true)
+                        ),
+                        $routeCollection->getResources()
+                    );
+                }
+            );
+
+            $nodes = include $configCache->getPath();
+            $this->treeCache->set($nodes);
         }
 
         return $nodes;
@@ -88,7 +126,7 @@ class RouteTree implements CacheClearerInterface, CacheWarmerInterface
      */
     public function clear ($cacheDir) : void
     {
-        $this->cache->clear();
+        $this->treeCache->clear();
     }
     //endregion
 
@@ -114,8 +152,30 @@ class RouteTree implements CacheClearerInterface, CacheWarmerInterface
      */
     public function warmUp ($cacheDir) : void
     {
-        $this->cache->clear();
+        $this->treeCache->clear();
         $this->generateItems();
     }
     //endregion
+
+
+    /**
+     * Creates and returns a new config cache
+     */
+    private function getConfigCacheFactory () : ConfigCacheFactoryInterface
+    {
+        if (null === $this->configCacheFactory)
+        {
+            $this->configCacheFactory = new ConfigCacheFactory($this->isDebug);
+        }
+
+        return $this->configCacheFactory;
+    }
+
+
+    /**
+     */
+    public function setConfigCacheFactory (ConfigCacheFactoryInterface $factory) : void
+    {
+        $this->configCacheFactory = $factory;
+    }
 }