| Q             | A
| ------------- | --------------------------------------------------------------------- |
| BC breaks?    | no                                                                |
| New feature?  | no <!-- don't forget to update CHANGELOG.md -->                   |
| Improvement?  | yes <!-- improves an existing feature, not adding a new one -->    |
| Bug fix?      | no                                                                |
| Deprecations? | no <!-- don't forget to update UPGRADE.md and CHANGELOG.md -->    |
| Docs PR       | — <!-- insert URL here -->                                  |

<!-- describe your changes below -->
So now the setup is like this:

1. `symfony/cache` as first level cache. Is disabled on debug
2. `ConfigCache` as second level cache.
3. Regenerate items

### Timing Differences

|        | dev (cache empty) | dev (cache filled) | prod (cache filled) |
| ------ | ----------------- | ------------------ | ------------------- |
| before | `40ms`            | (`40ms`) \*)       | `0.003ms`           |
| after  | `20ms`            | `5ms`              | `0.003ms`           |

\*) the cache was disabled in dev, so always "not filled"